refactor: ユニットテストをソースコードと同じディレクトリにコロケーション
ユニットテスト(19ファイル)を tests/ から src/ に移動し、
対応するソースファイルと同じディレクトリに配置。
VRT・統合テストは tests/ に残す。
